Applied patch from jira issue: OFBIZ-3277 - Problem in the createUpdateBillingAddress service.
From one page checkout if register user select shipping and billing address not same.
Go for continue then another Postal Address created in the system which should not be the case.
The new postal address should create only when user do some changes in the billing postal address (in case of register user and keep address book set to Y).

Thanks Deepak & Rishi for the contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@885702 13f79535-47bb-0310-9956-ffa450edef68