A patch from Sascha Rodekamp "sendMailFrom Screen don't need to load uiLabels" (https://issues.apache.org/jira/browse/OFBIZ-3969) - OFBIZ-3969
Sascha: I noticed an issue in the sendOrderNotificationScreen method. There is no need of loading uiLabels in the method, because the labels will be loaded in the screen (which will be used for the email). The point is, when someone (for whatever reason) don't add English uiLabels and the service try to access the default English labels the an exception will be thrown . This can be simply avoided.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1003836 13f79535-47bb-0310-9956-ffa450edef68