A modified patch from Anne Jessel "title-property should be titleProperty" (https://issues.apache.org/jira/browse/OFBIZ-4018) - OFBIZ-4018
Anne: some screen widgets set the field title-property instead of titleProperty. This causes errors in the logs, and means that the page title is not being set.
While preparing this patch, I discovered the labels referred to by some of these titleProperties were not available in Order component. However the missing ones were in Ecommerce. So I copied the missing ones to OrderUiLabels.xml.

JLR: instead of copying labels in OrderUiLabels.xml, I moved them to CommonUiLabels.xml: less and more consistent changes in the future.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1037559 13f79535-47bb-0310-9956-ffa450edef68