3 patches from Sascha Rodekamp "Localisation of labels in javascript" (https://issues.apache.org/jira/browse/OFBIZ-3964) - OFBIZ-3964
This cames from Adam's advice: http://markmail.org/message/kai5lfvlnsz7xwki

Sascha comes to this solution
The Idea is to have a method which takes a JSON Object with uiLabel Resources and label entries and returns the resolved labels in a JSON Object which can be used directly in the JS Code.
The advantage of this way, is that we have only one Ajax request which reads all labels. The alternative is an Ajax request for each label, which simplify the request (because you haven't to create a complex JSON Object) but is very expansive.

* Replaces all JS alerts with the new fancy error boxes
* Adds a few uiLabels

I simply fixed some typos. I did not test all cases but did a good review.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1056497 13f79535-47bb-0310-9956-ffa450edef68