A patch from Sergei Biletnikov  https://issues.apache.org/jira/browse/OFBIZ-4608 "Shopping cart update problem for several shipping groups"
That is pretty simple bug, but causes a big problems during the order editing in case if you have more than 1 shipping group, please, just look at the patch, it is very simple. I found the similar bug in the system already before, which was fixed successfully.

Also, this bug exists in the version 10.04, could you fix it there?


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1210211 13f79535-47bb-0310-9956-ffa450edef68