A slightly enhanced path from Sergei Biletnikov for "Search orders by good identification" https://issues.apache.org/jira/browse/OFBIZ-5031
I propose to extend the search mask for the search orders forms with the "Good identification" field.
A thing which can be improved is :
<option value="">Any good identification</option>   (in "applications/order/webapp/ordermgr/order/findOrders.ftl")
adding : "Any good identification" to the resource bundle.


Also, I found a bug which exists from 10.04 at least:
search orders by shipment method does not work properly, because:
{noformat}
applications/order/webapp/ordermgr/WEB-INF/actions/order/FindOrders.groovy

 currentCarrierShipmentMethod = delegator.findByAnd("CarrierShipmentMethod", [partyId : carrierPartyId, shipmentMethodTypeId : shipmentMethodTypeId], null, false);

is not correct, the result is the list, so the remedy is simple:

 currentCarrierShipmentMethod = EntityUtil.getFirst(delegator.findByAnd("CarrierShipmentMethod", [partyId : carrierPartyId, shipmentMethodTypeId : shipmentMethodTypeId], null, false));

{noformat}
My patch has this fix, but it is necessary to make this fix for other OFBiz version.

jleroux: I simply added the ProductAnyGoodIdentification label, I will commit the groovy fix apart for easier backport

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1493387 13f79535-47bb-0310-9956-ffa450edef68