A patch from János Zsolt for "Load cart from order loses OrderItemAttributes" https://issues.apache.org/jira/browse/OFBIZ-5286
When an order is created from a cart, OrderItemAttribute entities are created as expected. But when this same order is used to load back to the cart, the OrderItemAttributes are not reset. I found the cause of this issue in the org.ofbiz.order.shoppingcart.ShoppingCartServices.loadCartFromOrder(DispatchContext, Map<String, Object>) method.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1509164 13f79535-47bb-0310-9956-ffa450edef68