This fixes "Net before overhead report generates an error" https://issues.apache.org/jira/browse/OFBIZ-5267
This is related with r1361130, where the birt component was moved from framework to specialpurpose.

The fix is to create the possibility of a loose coupling from applications controllers to specialpurpose controllers. For that an if-present attribute is added to the include element. When used, if the included controller does not exist, instead of an error and a stack trace, a warning is thrown (only in English, for developers) suggesting to checkout the component from trunk. Indeed, in the current case, this situation should only arise in release after 12.04 where the birt component, like all others but the ecommerce component, had been removed. This can be used in other cases, but for now only the component level is covered.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1518777 13f79535-47bb-0310-9956-ffa450edef68