A patch from Christian Carlow for "Cancelled ship groups show incorrect quantities on edit order items page" https://issues.apache.org/jira/browse/OFBIZ-5405
Cancelled ship groups show incorrect quantities on the edit order items page.  I cancelled a ship group with a quantity of 2 for an order item but the quantity was still set to 2 when entering edit mode again.  Because its not set to 0 or blank, all cancelled ship groups must manually be set to one of those values in order for the quantity not to be added back after updating the order items.

This patch resolves the problem. It calculates the ship group quantity with OrderItemShipGroupAssoc.quantity - OrderItemShipGroupAssoc.cancelQuantity in editorderitems.ftl



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1550255 13f79535-47bb-0310-9956-ffa450edef68