From a wrong Christian Carlow's patch for "ShipBeforeDate/ShipAfterDate of additional ship groups displayed incorrectly on shipping options page of Customer Order Entry app" https://issues.apache.org/jira/browse/OFBIZ-5381
ShipBeforeDate/ShipAfterDate of ship groups are displayed incorrectly on the shipping options page of the Customer Order Entry app.  The dates is set correctly but the shipping options page does not get the shipping date by shipGroupIndex.  Instead it gets the shipBeforeDate/shipAfterDate of the order item.  The page should be changed to get the dates by passing in the shipGroupIndex: cart.getShipAfterDate(shipGroupIndex)

To reproduce
1.  Create new order for DemoCustCompany
2.  Add productId "GZ-BASKET" with quantity 10
3.  Click "Finalize Order" link
4.  Click "Create New Ship Group" button on Shipping page and then click "Continue" link
5.  Assign 5 of GZ-BASKET to Ship Group 2 and click "Continue" link
6.  Select shipping methods and set shipBeforeDate/shipAfterDate and click "Continue" link
7.  Click the "Options" link to return to previous ship options page
8.  Notice the dates of the second ship group are incorrectly set to the same values as the date of the first ship group

jleroux: Christian's patch was reversed, with his comment I eventually got the idea

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1552901 13f79535-47bb-0310-9956-ffa450edef68