Fix OFBIZ-5565: FinAccountHelper.getFinAccountFromCode() no longer returns financial account.
Now that EntityCondition lookups work against encrypted fields, there is
no longer a need to do the field encryption locally in the above method.
So, just pass the value directly into the entity engine.

There are also test cases, to prevent this breakage from occurring
again.

This commit depends on r1604967 and r1604968, at the minimum.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1604993 13f79535-47bb-0310-9956-ffa450edef68