Commit for https://issues.apache.org/jira/browse/OFBIZ-5843
Re-running the same load test with a service call proves that the problem was not in the service call but only in the minilang implementation. So putting back this part.

There are more work to do as suggested Scott on dev ML: the call could be done in the same transaction right before committing, to reduce the amount of time the lock is held. But for now it's enough

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1635567 13f79535-47bb-0310-9956-ffa450edef68