A patch from Bob Morley for "Data seeding to support ability to return products of type RAW_MATERIAL" https://issues.apache.org/jira/browse/OFBIZ-3669
This patch contains data seeding that supports the ability to have return items that contain products of type RAW_MATERIAL.  This allows an enterprise to return raw materials back to suppliers as well as customers returning them (if the enterprise sells raw materials to customers).  There is no code in this patch, just the data model wiring to map the raw material product types to the proper return types / accounting types.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1650949 13f79535-47bb-0310-9956-ffa450edef68