A patch from Pierre Smits for "Quote profit percentage is calculated wrong" https://issues.apache.org/jira/browse/OFBIZ-5858
When reviewing the profit page of a quote the profit percentage is calculated wrong. Instead of dividing the profit by the quote amount giving the margin, the quote amount is divided by the average cost (when provided). Thus giving the quote (amount) to cost ratio.

jleroux: there was an interesting discussion with Divesh Dutta who suggested to calculate the profit mark-up instead of the profit margin as done here.I agree we want the profit margin

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1661853 13f79535-47bb-0310-9956-ffa450edef68