Applied patch for OFBIZ-6526: ordermgr/control/searchorders findOrders service returns incorrect orderCount and therefore viewSize.
This patch replaces the HashSet with LinkedHashSet for fieldsToSelect because otherwise statusId was being added to the front of the set which was then passed to count distinct query causing incorrect counts. LinkedHashSet preserves the order in which items are added so orderId always appears at the front of the set for the count distinct query.

Thanks Christian Carlow for spotting this and providing the patch.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1704005 13f79535-47bb-0310-9956-ffa450edef68