An enhanced patch from Pierre Smits for "Order items for GZ-1006-5 via ecommerce/ordermgr shows error" https://issues.apache.org/jira/browse/OFBIZ-6643
When an order is entered via the ecommerce component for a GZ-1006-5 an error is shown in the order items section

jleroux: actually, as Vyom Jain suggested "This is happening due to changes OFBIZ-6276. Either r1704108 should be reverted or additional patches be provided to avoid this reported issue (which is only one of them). Removing deprecated delegator methods require thorough testing as many as used in scripts and templates." I checked there was only 3 such changes for findByAnd and no others but getRelated where called in *.ftl and *.groovy. This completes the findByAnd case I will create a new Jira for the rest (did not check already, 238 cases!)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1704478 13f79535-47bb-0310-9956-ffa450edef68