4th step for "Secure HTTP headers" https://issues.apache.org/jira/browse/OFBIZ-6766
The secure flag on cookies instructs the browser to only submit the cookie as part of requests over secure (HTTPS) connections. This prevents the cookie from being observed as plain text in transit over the network.
The HttpOnly flag instructs the browser that this cookie can only be accessed when sending an HTTP request. This prevents scripts running as part of a page from retrieving the value and is a defense against XSS attacks.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1719762 13f79535-47bb-0310-9956-ffa450edef68