An enhanced patch from Yashwant Dhakad based on Pierre Smits's for "Creating a QuoteRole doesn't return to overview of QuoteRoles" https://issues.apache.org/jira/browse/OFBIZ-6026
Currently, when adding a QuoteRole to a Quote the user stays in the screen associated with 'EditQuoteRole'. Thus the user has no confirmation that the QuoteRole is added successfully. It is better to show the overviews of the quote roles.

Yashwant Dhakad werified the patch and found that success message not shown on the screen, so updated the patch and now shown the success message when added role successfully.

jleroux: I moved the party field line to have the lookup show. Note: in case of error the message is still awful, not a big deal

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1746290 13f79535-47bb-0310-9956-ffa450edef68