Improved: Product based promotion calculation. (OFBIZ-7359)
In case of a promotion where X% discount has to be applied on a product, when more than one quantity of the product is added to order, respective multiple item adjustments are created. Improved the code to create a single adjustment for the total discount.

Thanks: Arvind Singh Tomar for contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/order@1765034 13f79535-47bb-0310-9956-ffa450edef68